Extract addwidget to widget
